Turn modifier keywords into annotations
This splits the bare minimal set of keywords required to understand the
structure of the code from the larger set of keywords that act as
modifiers.

The core idea is to put as much of the language on equal footing to code
users could write, while also allowing the modifiers to benefit from
all the facilities annotations usually have, like the ability to have
parameters and defaults, as well as allowing migration and deprecation.

This is a limited implementation, as all supported annotations are
hard-coded and no syntax for defining annotations in Dora exists.