Add {Int,Long,Float,Double,Bool}#toStringBinary
This change adds a helper function that's helpful when working with bits.

This is certainly not the most elegant, elaborate or flexible approach
(compared to languages that support arbitrary bases or compared to a
typeclass-based design that enables a maximum of abstraction), but it is
simple and straight-forward and should last until a better solution is
required.