lang: allow defining and calling nullary methods
# Why do this?

The idea is to allow decoupling how a member is implemented from how it is exposed,
allowing authors to reevaluate the costs of (re-)computation vs. storage
without breaking existing callers of such code.

This approach is based on experience with earlier attempts, like

- Java's "use getters/setters everywhere, just in case"
- C#'s "use properties, which are source incompatible with getters/setters and binary incompatible with fields"

and avoids the drawbacks these older designs.

As the distinction between `fun` and `let` is nevertheless important,
the idea is that an IDE should use different colors when referring to such members
(which provides more useful information than in Java's or C#'s case).

# What does this?

This change does not deliver on the binary compatibility improvement,
as the resolution of nullary `fun` vs. `let` happens too early;
but binary compatibility requires that this decision made much later,
i. e. at "link-time" when the runtime starts up and reads its bytecode.

The goal of this change is to allow experimenting with nullary methods in code,
without requiring this up-front investment.

When to use this?

The idea is that methods, which are "constant" for the instance they are called on,
are defined without ():

  true.not, 1.toString, array.length, iterator.hasNext
        vs.
  iterator.next(), random.nextInt(), ...