boots: Enable boots on most functions in stdlib
Now that bugs on arm64 are fixed as well, this commit
can be landed again.

Boots will now be used for all lazily compiled functions
with --always-boots. However, Cannon will still be used for
compiling boots itself and for the stdlib functions needed
for boots (so all methods compiled ahead-of-time).